Introduce null accounts as nano::account{nullptr}

+++ nano/lib/blocks.hpp
@@ -64,7 +64,7 @@ public:
 	bool deserialize (nano::stream &, nano::block_type);
 	static size_t size (nano::block_type);
 	nano::block_hash successor{ 0 };
-	nano::account account{ 0 };
+	nano::account account{ static_cast<std::uint64_t> (0) };
 	nano::amount balance{ 0 };
 	uint64_t height{ 0 };
 	uint64_t timestamp{ 0 };
